ENH: Better organization for more brain atlases
BrainConnMatrix: improved the organization of output connectivity atlas. Now it is able to run the same data with different brain atlas. This is useful to multiple analysis with different brain atlas.
connmatrixformat.m: Improved the selection of BEDPOSTX folders, which only selects the BrainConn folders
split_labels: Now the splitted brain labels are added in separate folders with the name of brain atlas that was used. In this way, it is better organized to a multiple brain atlas analysis.